vahe simonyan

+++ src/main/java/com/booking/bookingservice/dto/ScheduleDto.java
@@ -2,14 +2,16 @@ package com.booking.bookingservice.dto;
 
 import com.booking.bookingservice.model.Schedule;
 
+import java.math.BigDecimal;
+
 public record ScheduleDto(
   Long id,
   String name,
   int members,
-  Double price,
+  BigDecimal price,
   Schedule.ScheduleType scheduleType,
   Long propertyId,
   String startDate,
   String endDate,
-  int cost
+  BigDecimal cost
 ) {}
